cdash error
git-svn-id: https://svn.opendtect.org/od/trunk@38420 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiTools/uibasemap.cc
@@ -274,8 +274,8 @@ void uiBaseMap::setView( const uiWorldRect& wr )
 
 void uiBaseMap::updateTransform()
 {
-    const uiRect viewrect( 0, 0, view_.scene().width(),
-				 view_.scene().height() );
+    const uiRect viewrect( 0, 0, (int)view_.scene().width(),
+				 (int)view_.scene().height() );
 
     if ( mIsZero(wr_.width(),mDefEps) || mIsZero(wr_.height(),mDefEps) )
 	return;
